chore(observables): Added subscribe with observer, added lodash
