+++ modules/minimongo-observable/index.ts
@@ -1,3 +1,3 @@
 export * from "./observable-collection";
 export * from "./to-observable";
-export * from "./meteor-observable-call";
+export * from "./meteor-observable";
